Dungeon: Do prefab replacements for RandomMapObject & SpawnSyncedObject.
Network.GetNetworkPrefabFromName now lets you fetch a network prefab by the name of the prefab. This means, regardless of whether you're using decompiled assets or an empty stand in name of an asset. We can do a replacement on load & swap it to be the actual asset included with the engine